rename test cases to test results
