+++ tests/integration/fuzzing/fuzzing_test.py
@@ -1,10 +1,8 @@
 from pathlib import Path
 
-from protostar.commands.test.test_cases import PassedFuzzTestCaseResult
-from tests.integration.conftest import (
-    RunCairoTestRunnerFixture,
-    assert_cairo_test_cases,
-)
+from protostar.commands.test.test_results import PassedFuzzTestCaseResult
+from tests.integration.conftest import (RunCairoTestRunnerFixture,
+                                        assert_cairo_test_cases)
 
 
 async def test_basic(run_cairo_test_runner: RunCairoTestRunnerFixture):
